defines checksum module
supports reusing constants and types for checksums in multiple ironfish-frost
modules

we will likely use checksums to verify data in multiple dkg rounds
